+++ HISTORY.rst
@@ -4,6 +4,12 @@
 History
 -------
 
+0.2.3 (2015-10-18)
+------------------
+
+* Changed to a SHA-256 hash for the pseudo-unique transaction ID in order to improve
+  ID stability while preserving uniqueness.
+
 0.2.2 (2015-10-04)
 ------------------